requirements.txt
fonttools==4.53.1
langchain_core==0.3.28
langchain_openai==0.2.14
matplotlib==3.10.0
numpy==2.2.1
openai==1.58.1
Pillow==11.0.0
PyYAML==6.0.2
PyYAML==6.0.2
